+++ pilz_control/test/robot_mock.h
@@ -44,7 +44,7 @@ class RobotMock : public hardware_interface::RobotHW
 public:
   RobotMock();
 
-  void read();
+  std::array<JointData, NUM_JOINTS> read() const;
   void write(const ros::Duration& period);
 
   bool isMoving(const double& eps = JOINT_VELOCITY_EPS) const;